refactor : Session Token naming refactoring
