+++ src/main/java/com/flab/jbly/application/user/LoginService.java
@@ -2,6 +2,8 @@ package com.flab.jbly.application.user;
 
 import com.flab.jbly.application.user.command.LoginCommand;
 import com.flab.jbly.application.user.result.LoginResult;
+import com.flab.jbly.domain.auth.Session;
+import com.flab.jbly.domain.auth.SessionRepository;
 import com.flab.jbly.domain.user.PasswordEncryption;
 import com.flab.jbly.domain.user.UserRepository;
 import com.flab.jbly.infrastructure.exception.UserPasswordNotMatchedException;
@@ -10,6 +12,7 @@ import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -18,9 +21,11 @@ public class LoginService {
     private final PasswordEncryption encryption;
     private final UserRepository repository;
     private final HttpSession httpSession;
+    private final SessionRepository sessionRepository;
 
     Logger logger = LoggerFactory.getLogger(this.getClass());
 
+    @Transactional
     public LoginResult login(LoginCommand command) {
         var user = repository.findByUserId(command.userId());
 
@@ -29,6 +34,7 @@ public class LoginService {
         }
 
         httpSession.setAttribute("user",user);
+        sessionRepository.save(Session.of(user.getId(), httpSession.getId()));
         return new LoginResult(user.getUserId());
     }
 }
