+++ src/main/java/com/flab/jbly/domain/auth/Session.java
@@ -1,4 +1,4 @@
-package com.flab.jbly.domain.user;
+package com.flab.jbly.domain.auth;
 
 
 import jakarta.persistence.Column;
@@ -9,10 +9,12 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "session")
+@Table(name = "token")
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Session {
