+++ src/main/java/com/flab/jbly/domain/auth/SessionRepository.java
@@ -1,6 +1,8 @@
-package com.flab.jbly.domain.user;
+package com.flab.jbly.domain.auth;
 
 public interface SessionRepository {
 
     void save(Session session);
+
+    Session findById(Long id);
 }
