+++ src/main/java/com/flab/jbly/infrastructure/auth/SessionJpaRepository.java
@@ -1,6 +1,6 @@
 package com.flab.jbly.infrastructure.auth;
 
-import com.flab.jbly.domain.user.Session;
+import com.flab.jbly.domain.auth.Session;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SessionJpaRepository extends JpaRepository<Session, Long> {
