+++ src/main/java/com/flab/jbly/presentation/user/LoginController.java
@@ -4,6 +4,7 @@ import com.flab.jbly.application.user.LoginService;
 import com.flab.jbly.application.user.result.LoginResult;
 import com.flab.jbly.presentation.user.request.LoginRequest;
 import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,5 +21,16 @@ public class LoginController {
 	public LoginResult login(@RequestBody LoginRequest request) {
 		var result = service.login(request.toCommand());
 		return result;
+	}
+
+	@GetMapping("test")
+	public String test() {
+		return "Hello";
+	}
+
+	// Session Redirect
+	@GetMapping("/sessionTest")
+	public String sessionTest() {
+		return "it's working";
 	}
 }
